Create all my initial models.
This is entirely for the domain. Remember, I had previously built a prototype,
so I already have certain schema restrictions, as I did the design work
previously.

I also have two diagrams of the schema in public. The second one is the real
one, after I ran the generators, I realized that stuff was a bit weird.

Here are the commands that I ran to generate these:

    $ bin/rails g model shop name:string

    $ bin/rails g model customer name:string phone:string email:string \
    >  date_on_waiting_list:datetime status:integer local:boolean
    >  location:string credit:string want:text notes:text shop:references

    $ bin/rails generate model appointment date_scheduled:datetime
    >  description:text customer:references artist:references

    $ bin/rails g model artist name:string

Afterwards, I deleted the test files. I'm always torn between generators and
just doing it by hand, but I like to have generators make the migrations, so in
this case I chose to use the generators. With a controller or view, I may not.

This is the first commit which has a migration, and so I'm curious to see what
happens when I push to Heroku. I bet that I should be telling it to run
`bin/rake db:migrate` after each deploy. But maybe it does that already? I'm
not sure. Only one way to find out! I can afford to be sloppy like this now,
since I have no users. :wink: