Tell Heroku to migrate after deploy
Yup! My hunch in 8168f626 was correct. While it didn't take the site down,
that's just because I didn't have any pages which interacted with models. I
figured this out by:

- letting it deploy:
  https://travis-ci.org/artisan-tattoo/artisan_assistant/builds/27576375
- entering the console (`heroku run rails console`)
- loading up a Shop (`Shop.new`)

With this, it threw a big stack trace about not being able to do that. Great. Now I know.

As detailed in Travis' docs[1], you can use this syntax to have it run commands
after you deploy. So we'll migrate the database and then restart.

1: http://docs.travis-ci.com/user/deployment/heroku/