Incorrect email leads to sign in failure.
While this just adds a test, I _did_ comment out the implementation for
the previous test, just to watch it fail. It just so happens that either
causes the same failure. It's nice to have a test for both, though.

These tests are probably overkill for something that's not as important
as signing in. I want to make 100% sure this works all the time, though,
so I'll expend the effort to make it overly tested.