Add bootstrap.
I'm using the bootstrap-sass gem since it's now officially part of Bootstrap.
This is generally how you're supposed to use libraries with the asset pipeline:
add a gem which takes care of updating things for you and doing the
integration. I'm not 100% sold on it, exactly, because it relies on people
doing the wrapping rather than importing them directly, but I don't plan on
doing a whole lot of fancy in this app anyway, so it shouldn't be an issue.

The gem instructions don't tell you about adding the file; it took me a bit of
searching to realize that the gem only made the sass files available, rather
than adding bootstrap itself to your project. This makes sense, I guess, if
you want to pick and choose which parts of Bootstrap get used, or to only
use them on part of your sites.