New home page.
Here's a trick that I like in Rails apps: you basically just render the
homepage with `layout: false`, and then shove everything into the single view.

This is an example of where DRY can lead you astray. Yes, I copy and pasted a
bunch of stuff into the view from a layout. Yes, I could, in theory, extract a
common layout and use it for both. No, that would not be worth it. Here's why:

The home page is often just marketing material, and ends up being significantly
different than the rest of the site. Depending on how big your organization is,
it can also end up changing a lot for various reasons, too.

Using the common layout doesn't actually buy you anything, because _the home
page and the rest of your site don't actually have a lot in common_. The
purpose of DRY is to make future change easier. Trying to have this page share
a layout with the rest of the site actually makes future change harder, in my
experience.

Eventually, when I'm ready to launch, this won't be a Bootstrap template, and
will be something real instead.