Fix up testing artist/shop association
I wasn't happy with this test, and in
https://github.com/artisan-tattoo/artisan_assistant/commit/423e50b46b1f88e6bf2bfec2bdbd7e57bfcdb1eb
we had a discussion about it. @qrush was correct that the way I was testing
this was much more of a model level concern.

I also considered making this a controller test, but I tend to lean a little
bit feature-heavy in my testing, and this is what I actually care about: that
the artist properly appears on the dashboard after creation.